pyproject.toml
[project]
name = "photo-analysis"
version = "0.1.0"
description = "Analysis for photographs I took"
authors = [
    {name = "Justin-Qian",email = "yjqian19@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
